# openapi.yaml
openapi: 3.0.3
info:
  title: AgentX Registry API
  version: 0.8.0
  description: |
    Hosted AgentX API. Business routes require a bearer token and return a stable
    ErrorEnvelope on JSON errors. Unknown dependency failures are logged with the
    request ID and return stable non-sensitive messages. Workspace roles are
    hierarchical: viewer, developer, admin, owner. Legacy unscoped routes are
    disabled in hosted mode. Device and release creation require idempotency keys.
    Expected duplicate and missing-resource outcomes use stable public 4xx envelopes;
    unexpected repository and dependency failures remain non-sensitive 500 responses.
security:
  - bearerAuth: []
servers:
  - url: https://{host}
    description: Hosted deployment
    variables:
      host: { default: registry.example.com }
paths:
...
  /v1/auth/config:
    get:
      operationId: getOIDCClientConfig
      summary: Public OIDC Device Authorization configuration for the CLI
      description: Returns public issuer, client ID, audience, and scope values; never returns a client secret.
      security: []
      responses:
        '200':
          description: OIDC Device Authorization configuration
          content:
            application/json: { schema: { '$ref': '#/components/schemas/OIDCClientConfig' } }
        '404': { '$ref': '#/components/responses/NotFound' }
  /v1/site/config:
    get:
      operationId: getPublicSiteConfig
      summary: Public legal and operational contact configuration
      description: Returns only owner-supplied public URLs and staffed contact addresses. Hosted startup fails when any value is absent.
      security: []
      responses:
        '200': { description: Public site configuration, content: { application/json: { schema: { '$ref': '#/components/schemas/SiteConfig' } } } }
        '404': { '$ref': '#/components/responses/NotFound' }
...
  /v1/invitations:
    get:
      operationId: listMyWorkspaceInvitations
      summary: List pending invitations for the authenticated verified email
      description: Requires an explicit true email_verified OIDC claim. Only non-expired pending invitations for the canonical authenticated email are returned.
      parameters: [{ '$ref': '#/components/parameters/Limit' }, { '$ref': '#/components/parameters/Cursor' }]
      responses:
        '200': { '$ref': '#/components/responses/InvitationPage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
  /v1/invitations/{invitation_id}/claim:
    post:
      operationId: claimWorkspaceInvitation
      summary: Claim a pending invitation
      description: Atomically creates membership only when the authenticated principal has an explicit true email_verified claim matching the invitation email.
      parameters: [{ '$ref': '#/components/parameters/InvitationID' }]
      responses:
        '200': { description: Claimed invitation, content: { application/json: { schema: { '$ref': '#/components/schemas/WorkspaceInvitation' } } } }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
        '404': { '$ref': '#/components/responses/NotFound' }
  /v1/workspaces:
    get:
      operationId: listWorkspaces
      summary: List workspaces visible to the principal
      parameters: [{ '$ref': '#/components/parameters/Limit' }, { '$ref': '#/components/parameters/Cursor' }]
      responses:
        '200': { '$ref': '#/components/responses/WorkspacePage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '401': { '$ref': '#/components/responses/Unauthorized' }
    post:
      operationId: createWorkspace
      summary: Create a workspace and owner membership
      description: A duplicate slug returns a public 400 conflict envelope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string, minLength: 1 }
                slug: { type: string, pattern: '^[a-z0-9][a-z0-9-]{1,62}$' }
      responses:
        '201': { description: Created workspace, content: { application/json: { schema: { '$ref': '#/components/schemas/Workspace' } } } }
        '400': { '$ref': '#/components/responses/BadRequest' }
  /v1/workspaces/{id}:
    delete:
      operationId: deleteWorkspace
      summary: Delete a workspace
      description: Requires owner. The database deletion, retained tombstone audit, and one durable cleanup event per candidate artifact commit atomically; the worker deletes only globally unreferenced digests. An active platform legal hold blocks deletion.
      x-agentx-minimum-role: owner
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }]
      responses:
        '204': { description: Deleted }
        '403': { '$ref': '#/components/responses/Forbidden' }
        '404': { '$ref': '#/components/responses/NotFound' }
  /v1/workspaces/{id}/members:
    get:
      operationId: listWorkspaceMembers
      summary: List workspace memberships
      description: Requires viewer.
      x-agentx-minimum-role: viewer
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }, { '$ref': '#/components/parameters/Limit' }, { '$ref': '#/components/parameters/Cursor' }]
      responses:
        '200': { '$ref': '#/components/responses/MembershipPage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
  /v1/workspaces/{id}/invitations:
    get:
      operationId: listWorkspaceInvitations
      summary: List workspace invitations
      description: Requires admin. Includes pending, accepted, revoked, and expired invitation records.
      x-agentx-minimum-role: admin
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }, { '$ref': '#/components/parameters/Limit' }, { '$ref': '#/components/parameters/Cursor' }]
      responses:
        '200': { '$ref': '#/components/responses/InvitationPage' }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
    post:
      operationId: createWorkspaceInvitation
      summary: Create an expiring workspace invitation
      description: Requires admin. Email is canonicalized and the role cannot be owner. Lifetime defaults to seven days and must be between 15 minutes and 30 days.
      x-agentx-minimum-role: admin
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }]
      requestBody:
        required: true
        content:
          application/json: { schema: { '$ref': '#/components/schemas/InvitationCreate' } }
      responses:
        '201': { description: Created invitation, content: { application/json: { schema: { '$ref': '#/components/schemas/WorkspaceInvitation' } } } }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
  /v1/workspaces/{id}/invitations/{invitation_id}:
    delete:
      operationId: revokeWorkspaceInvitation
      summary: Revoke a pending workspace invitation
      description: Requires admin. Accepted, expired, and already revoked invitations cannot be revoked.
      x-agentx-minimum-role: admin
      parameters: [{ '$ref': '#/components/parameters/WorkspaceID' }, { '$ref': '#/components/parameters/InvitationID' }]
      responses:
        '204': { description: Invitation revoked }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
        '404': { '$ref': '#/components/responses/NotFound' }
  /v1/workspaces/{id}/members/{user_id}:
    parameters:
      - { '$ref': '#/components/parameters/WorkspaceID' }
      - { name: user_id, in: path, required: true, schema: { type: string } }
    patch:
      operationId: updateWorkspaceMemberRole
      summary: Change a member role
      description: Requires admin. Owner membership cannot be changed.
      x-agentx-minimum-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties: { role: { '$ref': '#/components/schemas/AssignableRole' } }
      responses:
        '204': { description: Role updated }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
        '404': { '$ref': '#/components/responses/NotFound' }
    delete:
      operationId: removeWorkspaceMember
      summary: Remove a member
      description: Requires admin. Owner membership cannot be removed.
      x-agentx-minimum-role: admin
      responses:
        '204': { description: Member removed }
        '400': { '$ref': '#/components/responses/BadRequest' }
        '403': { '$ref': '#/components/responses/Forbidden' }
        '404': { '$ref': '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: OIDC JWT }
  parameters:
    WorkspaceID: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    Limit: { name: limit, in: query, required: false, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
    Cursor: { name: cursor, in: query, required: false, description: Opaque value from next_cursor or X-Next-Cursor, schema: { type: string } }
...
    InvitationID: { name: invitation_id, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    WorkspacePage: { description: Workspace page, content: { application/json: { schema: { allOf: [{ '$ref': '#/components/schemas/Page' }, { type: object, properties: { items: { type: array, items: { '$ref': '#/components/schemas/Workspace' } } } }] } } } }
    MembershipPage: { description: Membership page, content: { application/json: { schema: { allOf: [{ '$ref': '#/components/schemas/Page' }, { type: object, properties: { items: { type: array, items: { '$ref': '#/components/schemas/Membership' } } } }] } } } }
    InvitationPage: { description: Workspace invitation page, content: { application/json: { schema: { allOf: [{ '$ref': '#/components/schemas/Page' }, { type: object, properties: { items: { type: array, items: { '$ref': '#/components/schemas/WorkspaceInvitation' } } } }] } } } }
...
    BadRequest: { description: Invalid request, content: { application/json: { schema: { '$ref': '#/components/schemas/ErrorEnvelope' } } } }
    Unauthorized: { description: Missing or invalid bearer token, content: { application/json: { schema: { '$ref': '#/components/schemas/ErrorEnvelope' } } } }
    Forbidden: { description: Workspace membership or role denied, content: { application/json: { schema: { '$ref': '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { '$ref': '#/components/schemas/ErrorEnvelope' } } } }
...
  schemas:
    OIDCClientConfig:
      type: object
      required: [issuer, client_id, scope]
      properties:
        issuer: { type: string, format: uri }
        client_id: { type: string, minLength: 1 }
        audience: { type: string }
        scope: { type: string, minLength: 1 }
    SiteConfig:
      type: object
      required: [terms_url, privacy_url, support_url, abuse_email, security_email]
      properties:
        terms_url: { type: string, format: uri }
        privacy_url: { type: string, format: uri }
        support_url: { type: string, format: uri }
        abuse_email: { type: string, format: email }
        security_email: { type: string, format: email }
    Page:
      type: object
      required: [items, count]
      properties:
        items: { type: array, items: {} }
        count: { type: integer, minimum: 0, description: Total matching items before page slicing }
        next_cursor: { type: string }
...
    Workspace:
      type: object
      required: [id, slug, name, created_at]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        created_at: { type: string, format: date-time }
    Role: { type: string, enum: [viewer, developer, admin, owner] }
    AssignableRole: { type: string, enum: [viewer, developer, admin] }
    InvitationCreate:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        role: { '$ref': '#/components/schemas/AssignableRole' }
        expires_in_seconds: { type: integer, format: int64, minimum: 900, maximum: 2592000, default: 604800 }
    Membership:
      type: object
      required: [workspace_id, user_id, role, created_at]
      properties:
        workspace_id: { type: string, format: uuid }
        user_id: { type: string }
        role: { '$ref': '#/components/schemas/Role' }
        created_at: { type: string, format: date-time }
    WorkspaceInvitation:
      type: object
      required: [id, workspace_id, email, role, created_by, created_at, expires_at, status]
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        workspace_name: { type: string }
        workspace_slug: { type: string }
        email: { type: string, format: email }
        role: { '$ref': '#/components/schemas/AssignableRole' }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        revoked_at: { type: string, format: date-time }
        revoked_by: { type: string }
        accepted_at: { type: string, format: date-time }
        accepted_by: { type: string }
        status: { type: string, enum: [pending, accepted, revoked, expired] }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string, minLength: 1 }